Railway/requirements.txt
attrdict==2.0.1
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cytoolz==0.9.0.1
Django==2.2.3
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.1
hexbytes==0.1.0
idna==2.8
infura==0.2.1
itsdangerous==1.1.0
Jinja2==2.10.1
lru-dict==1.1.6
MarkupSafe==1.1.1
parsimonious==0.8.1
py-solc-x==0.4.0
pycparser==2.19
pycryptodome==3.8.2
pytz==2019.1
requests==2.22.0
requests-cache==0.5.0
rlp==1.1.0
semantic-version==2.6.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.25.3
web3==4.9.2
websockets==6.0
Werkzeug==0.15.4
attrdict==2.0.1
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cytoolz==0.9.0.1
Django==2.2.3
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.1
gunicorn==19.9.0
hexbytes==0.1.0
idna==2.8
infura==0.2.1
itsdangerous==1.1.0
Jinja2==2.10.1
lru-dict==1.1.6
MarkupSafe==1.1.1
parsimonious==0.8.1
pkg-resources==0.0.0
py-solc-x==0.4.0
pycparser==2.19
pycryptodome==3.8.2
pytz==2019.1
requests==2.22.0
requests-cache==0.5.0
rlp==1.1.0
semantic-version==2.6.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.25.3
web3==4.9.2
websockets==6.0
Werkzeug==0.15.4
